+++ Robot/src/main/java/com/js/dawa/model/arene/CaseArene.java
@@ -7,9 +7,9 @@ import java.util.Map;
 
 import com.js.dawa.iu.arene.render.CaseAreneRenderDefaut;
 import com.js.dawa.iu.arene.render.CaseRender;
+import com.js.dawa.model.position.Position;
 import com.js.dawa.model.robot.Attribut;
 import com.js.dawa.model.robot.DataBoard;
-import com.js.dawa.model.robot.Position;
 
 public class CaseArene implements ObjetArene{
 	
@@ -70,7 +70,7 @@ public class CaseArene implements ObjetArene{
 	}
 
 	@Override
-	public void add(int px, int py) {
+	public void add(double px, double py) {
 		//objet immobile
 		
 	}
