+++ Robot/src/main/java/com/js/dawa/model/arene/Mine.java
@@ -9,9 +9,9 @@ import com.js.dawa.iu.arene.render.CaseRender;
 import com.js.dawa.iu.arene.render.FireBallRender;
 import com.js.dawa.iu.arene.render.InfoRender;
 import com.js.dawa.iu.arene.render.HurtObjetRender;
+import com.js.dawa.model.position.Position;
 import com.js.dawa.model.robot.Attribut;
 import com.js.dawa.model.robot.DataBoard;
-import com.js.dawa.model.robot.Position;
 
 public class Mine implements ObjetArene {
 
@@ -82,7 +82,7 @@ public class Mine implements ObjetArene {
 	}
 
 	@Override
-	public void add(int px, int py) {
+	public void add(double px, double py) {
 		//na : mine is immobile
 
 	}
